+++ R/ab_property.R
@@ -241,20 +241,7 @@ ab_ddd <- function(x, administration = "oral", ...) {
   meet_criteria(administration, is_in = c("oral", "iv"), has_length = 1)
 
   x <- as.ab(x, ...)
-  ddd_prop <- administration
-  # old behaviour
-  units <- list(...)$units
-  if (!is.null(units) && isTRUE(units)) {
-    if (message_not_thrown_before("ab_ddd", entire_session = TRUE)) {
-      warning_(
-        "in `ab_ddd()`: using `ab_ddd(..., units = TRUE)` is deprecated, use `ab_ddd_units()` to retrieve units instead.",
-        "This warning will be shown once per session."
-      )
-    }
-    ddd_prop <- paste0(ddd_prop, "_units")
-  } else {
-    ddd_prop <- paste0(ddd_prop, "_ddd")
-  }
+  ddd_prop <- paste0(administration, "_ddd")
   out <- ab_validate(x = x, property = ddd_prop)
 
   if (any(ab_name(x, language = NULL) %like% "/" & is.na(out))) {
@@ -274,16 +261,17 @@ ab_ddd_units <- function(x, administration = "oral", ...) {
   meet_criteria(administration, is_in = c("oral", "iv"), has_length = 1)
 
   x <- as.ab(x, ...)
-  if (any(ab_name(x, language = NULL) %like% "/")) {
+  ddd_prop <- paste0(administration, "_units")
+  out <- ab_validate(x = x, property = ddd_prop)
+
+  if (any(ab_name(x, language = NULL) %like% "/" & is.na(out))) {
     warning_(
-      "in `ab_ddd_units()`: DDDs of combined products are available for different dose combinations and not (yet) part of the AMR package.",
+      "in `ab_ddd_units()`: DDDs of some combined products are available for different dose combinations and not (yet) part of the AMR package.",
       "Please refer to the WHOCC website:\n",
       "www.whocc.no/ddd/list_of_ddds_combined_products/"
     )
   }
-
-  ddd_prop <- paste0(administration, "_units")
-  ab_validate(x = x, property = ddd_prop)
+  out
 }
 
 #' @rdname ab_property